use exc_handle-auth-files-form-header-cookie-cors

+++ db/db_user.py
@@ -1,5 +1,5 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
-
 from db.hash import Hash
 from db.models import DbUser
 from db.schemas import UserBase
@@ -23,12 +23,24 @@ def get_all_user(db:Session):
     return db.query(DbUser).all()
 
 def get_user(db:Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail= f"User with id {id} not found !")
+    return user
+
+def get_user_by_username(db:Session, username: str):
+    user = db.query(DbUser).filter(DbUser.username == username).first()
+    if not user:
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail= f"User with username {username} not found !")
+    return user
 
 # Update User
 
 def update_user(db:Session, id:int, request:UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail= f"User with id {id} not found !")
+    
     user.update({
         DbUser.username : request.username,
         DbUser.email : request.email,
@@ -41,6 +53,8 @@ def update_user(db:Session, id:int, request:UserBase):
 
 def delete_user(db:Session, id:int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail= f"User with id {id} not found !")
     db.delete(user)
     db.commit()
